remove data resolver APIs
